feat: add new chart that aligns with DPF contract
Move from multi chart to a single chart + resource injector as separate
chart to be able to make use of the serviceDaemonSet variables correctly
in the relevant resources.

The old chart remains as is, but can be removed after we start using
this chart, up to the team to decide.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>